Zscore Updates
Added code and modified code to account for z-scored data
